refactor: create backend package skeleton, move models and agents (#303)
Steps 1-3 of the semantic graph module extraction: create the
backend/ package structure, move models/ to backend/model/ and
agents/ to backend/agents/ as the canonical sources, and convert
the original packages into re-export shims for backward compatibility.
