+++ backend/__init__.py
@@ -1,0 +1,1 @@
+"""Backend domain logic — models, agents, and semantic graph pipeline."""
