+++ backend/model/__init__.py
@@ -1,0 +1,32 @@
+"""Pydantic models used as ``output_type`` by agents.
+
+The canonical source for model definitions. The top-level ``models/``
+package is a re-export shim that will be removed once all callers
+migrate to ``backend.model``.
+"""
+
+from .semantic_graph import (
+    Classification,
+    ClassificationKind,
+    EdgeRole,
+    EdgeSemantic,
+    Enrichment,
+    NodeType,
+    Role,
+    SemanticGraph,
+    SemanticGraphEdge,
+    SemanticGraphNode,
+)
+
+__all__ = [
+    "Classification",
+    "ClassificationKind",
+    "EdgeRole",
+    "EdgeSemantic",
+    "Enrichment",
+    "NodeType",
+    "Role",
+    "SemanticGraph",
+    "SemanticGraphEdge",
+    "SemanticGraphNode",
+]
